+++ virtualizarr/readers/dmrpp.py
@@ -8,6 +8,7 @@ import numpy as np
 from xarray import Coordinates, Dataset, Index, Variable
 
 from virtualizarr.manifests import ChunkManifest, ManifestArray
+from virtualizarr.manifests.manifest import validate_and_normalize_path_to_uri
 from virtualizarr.readers.common import VirtualBackend
 from virtualizarr.types import ChunkKey
 from virtualizarr.utils import _FsspecFSFromFilepath, check_for_collisions
@@ -23,6 +24,7 @@ class DMRPPVirtualBackend(VirtualBackend):
         loadable_variables: Iterable[str] | None = None,
         decode_times: bool | None = None,
         indexes: Mapping[str, Index] | None = None,
+        virtual_backend_kwargs: Optional[dict] = None,
         reader_options: Optional[dict] = None,
     ) -> Dataset:
         loadable_variables, drop_variables = check_for_collisions(
@@ -30,10 +32,19 @@ class DMRPPVirtualBackend(VirtualBackend):
             loadable_variables=loadable_variables,
         )
 
+        if virtual_backend_kwargs:
+            raise NotImplementedError(
+                "DMR++ reader does not understand any virtual_backend_kwargs"
+            )
+
         if loadable_variables != [] or decode_times or indexes is None:
             raise NotImplementedError(
                 "Specifying `loadable_variables` or auto-creating indexes with `indexes=None` is not supported for dmrpp files."
             )
+
+        filepath = validate_and_normalize_path_to_uri(
+            filepath, fs_root=Path.cwd().as_uri()
+        )
 
         fpath = _FsspecFSFromFilepath(
             filepath=filepath, reader_options=reader_options
@@ -84,6 +95,8 @@ class DMRParser:
     _DEFAULT_ZLIB_VALUE = 6
     # Encoding keys that should be removed from attributes and placed in xarray encoding dict
     _ENCODING_KEYS = {"_FillValue", "missing_value", "scale_factor", "add_offset"}
+    root: ET.Element
+    data_filepath: str
 
     def __init__(self, root: ET.Element, data_filepath: Optional[str] = None):
         """
@@ -91,9 +104,8 @@ class DMRParser:
 
         Parameters
         ----------
-        dmrpp_str : str
-            The dmrpp file contents as a string.
-
+        root: xml.ElementTree.Element
+            Root of the xml tree struture of a DMR++ file.
         data_filepath : str, optional
             The path to the actual data file that will be set in the chunk manifests.
             If None, the data file path is taken from the DMR++ file.